+++ src/models/operations/createinputsystembypack.ts
@@ -6056,6 +6056,38 @@ export const CreateInputSystemByPackAuthenticationMethodMicrosoftGraph = {
 export type CreateInputSystemByPackAuthenticationMethodMicrosoftGraph =
   OpenEnum<typeof CreateInputSystemByPackAuthenticationMethodMicrosoftGraph>;
 
+/**
+ * Microsoft 365 subscription plan for your organization, typically Microsoft 365 Enterprise
+ */
+export const CreateInputSystemByPackSubscriptionPlan = {
+  /**
+   * Microsoft 365 Enterprise
+   */
+  EnterpriseGcc: "enterprise_gcc",
+  /**
+   * Microsoft 365 GCC
+   */
+  Gcc: "gcc",
+  /**
+   * Microsoft 365 GCC High
+   */
+  GccHigh: "gcc_high",
+  /**
+   * Microsoft 365 DoD
+   */
+  Dod: "dod",
+  /**
+   * Microsoft 365 China (21Vianet)
+   */
+  China: "china",
+} as const;
+/**
+ * Microsoft 365 subscription plan for your organization, typically Microsoft 365 Enterprise
+ */
+export type CreateInputSystemByPackSubscriptionPlan = OpenEnum<
+  typeof CreateInputSystemByPackSubscriptionPlan
+>;
+
 export type CreateInputSystemByPackInputMicrosoftGraph = {
   /**
    * Unique ID for this input
@@ -6179,7 +6211,7 @@ export type CreateInputSystemByPackInputMicrosoftGraph = {
   /**
    * Microsoft 365 subscription plan for your organization, typically Microsoft 365 Enterprise
    */
-  planType?: models.SubscriptionPlanOptions | undefined;
+  planType?: CreateInputSystemByPackSubscriptionPlan | undefined;
   /**
    * Select or create a secret that references your client_secret to pass in the OAuth request parameter.
    */
@@ -15099,6 +15131,13 @@ export const CreateInputSystemByPackAuthenticationMethodMicrosoftGraph$outboundS
   );
 
 /** @internal */
+export const CreateInputSystemByPackSubscriptionPlan$outboundSchema: z.ZodType<
+  string,
+  z.ZodTypeDef,
+  CreateInputSystemByPackSubscriptionPlan
+> = openEnums.outboundSchema(CreateInputSystemByPackSubscriptionPlan);
+
+/** @internal */
 export type CreateInputSystemByPackInputMicrosoftGraph$Outbound = {
   id: string;
   type: "microsoft_graph";
@@ -15185,7 +15224,7 @@ export const CreateInputSystemByPackInputMicrosoftGraph$outboundSchema:
     tenantId: z.string().optional(),
     clientId: z.string().optional(),
     resource: z.string().optional(),
-    planType: models.SubscriptionPlanOptions$outboundSchema.optional(),
+    planType: CreateInputSystemByPackSubscriptionPlan$outboundSchema.optional(),
     textSecret: z.string().optional(),
     certOptions: models.CertOptionsType$outboundSchema.optional(),
     __template_url: z.string().optional(),