+++ RELEASES.md
@@ -1459,3 +1459,13 @@ Based on:
 - [typescript v0.7.0-alpha.14] .
 ### Releases
 - [NPM v0.7.0-alpha.14] https://www.npmjs.com/package/cribl-control-plane/v/0.7.0-alpha.14 - .
+
+## 2026-04-04 02:11:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.7.0-alpha.15] .
+### Releases
+- [NPM v0.7.0-alpha.15] https://www.npmjs.com/package/cribl-control-plane/v/0.7.0-alpha.15 - .